+++ lib/pantheios/services/coloured_console_log_service.rb
@@ -1,11 +1,11 @@
 
 # ######################################################################## #
-# File:         lib/pantheios/services/simple_console_log_service.rb
+# File:         lib/pantheios/services/coloured_console_log_service.rb
 #
 # Purpose:      Definition of the
-#               ::Pantheios::Services::SimpleConsoleLogService class
+#               ::Pantheios::Services::ColouredConsoleLogService class
 #
-# Created:      14th June 2015
+# Created:      19th June 2019
 # Updated:      4th June 2020
 #
 # Home:         http://github.com/synesissoftware/Pantheios-Ruby
@@ -13,7 +13,7 @@
 # Author:       Matthew Wilson
 #
 # Copyright (c) 2019-2020, Matthew Wilson and Synesis Information Systems
-# Copyright (c) 2015-2018, Matthew Wilson and Synesis Software
+# Copyright (c) 2019, Matthew Wilson and Synesis Software
 # All rights reserved.
 #
 # Redistribution and use in source and binary forms, with or without
@@ -46,6 +46,8 @@
 # ######################################################################## #
 
 
+require 'pantheios/services/common/console'
+
 =begin
 =end
 
@@ -58,7 +60,24 @@ module Services
 # NOTE: The *LogService* protocol is implemented by a class that provides
 # the instance methods +severity_logged?(severity : Object) : boolean+ and
 # +log(severity : Object, t : Time, prefix : String|Array, msg : String)+
-class SimpleConsoleLogService
+class ColouredConsoleLogService
+
+	module Constants
+
+		include ::Pantheios::Services::Common::Console::AnsiEscapeSequences
+	end # module Constants
+
+	def self.requires_prefix?
+
+		return @requires_prefix unless @requires_prefix.nil?
+
+		@requires_prefix = ::Pantheios::Services::Common::Console::Internal_::SHOULD_COLOURIZE_ ? :parts : false
+	end
+
+	def requires_prefix?
+
+		self.class.requires_prefix?
+	end
 
 	def severity_logged? severity
 
@@ -68,6 +87,98 @@ class SimpleConsoleLogService
 	def log sev, t, pref, msg
 
 		stm = infer_stream sev
+
+		if requires_prefix?
+
+			pref = pref.map do |part|
+
+				bg	=	Constants::Background
+				fg	=	Constants::Foreground
+
+				if part.respond_to?(:severity)
+
+					part = fg.bold part
+
+					case sev
+					when :violation
+
+						part = bg.red part
+						#part = fg.bright_magenta part
+						part = fg.bright_yellow part
+						part = fg.blinking part
+					when :alert
+
+						part = bg.red part
+						part = fg.bright_cyan part
+						part = fg.blinking part
+					when :critical
+
+						part = bg.red part
+						part = fg.white part
+					when :failure
+
+						part = bg.yellow part
+						part = fg.red part
+					when :warning
+
+						part = bg.yellow part
+						part = fg.blue part
+					when :notice
+
+						part = bg.dark_grey part
+						part = fg.white part
+					when :informational
+
+						part = bg.dark_grey part
+						part = fg.light_grey part
+					when :debug0
+
+						part = bg.blue part
+						part = fg.light_grey part
+					when :debug1
+
+						part = bg.blue part
+						part = fg.light_grey part
+					when :debug2
+
+						part = bg.blue part
+						part = fg.light_grey part
+					when :debug3
+
+						part = bg.blue part
+						part = fg.light_grey part
+					when :debug4
+
+						part = bg.blue part
+						part = fg.light_grey part
+					when :debug5
+
+						part = bg.blue part
+						part = fg.light_grey part
+					when :trace
+
+						part = bg.blue part
+						part = fg.light_grey part
+					when :benchmark
+
+						part = bg.black part
+						part = fg.light_grey part
+					else
+
+						;
+					end
+				else
+
+					part = fg.dark_grey part
+				end
+
+				part
+			end.join(', ')
+
+			pref = '[' + pref + ']: '
+
+			#pref = pref.map { |pp| pp.severity? ? map_sev_(sev) : sev }.join(
+		end
 
 		stm.puts "#{pref}#{msg}"
 	end
@@ -81,7 +192,9 @@ class SimpleConsoleLogService
 
 		$stderr
 	end
-end # class SimpleConsoleLogService
+
+	private
+end # class ColouredConsoleLogService
 
 end # module Services
 end # module Pantheios